+++ LabelStoreMax/CHANGELOG.md
@@ -1,9 +1,10 @@
 ## [5.7.1] - 2022-02-07
 
 * Refactor account order detail page
-* Fix continuous loading if users has no orders
+* Fix continuous loading if a user has no orders
 * New styling for tabs in the account order detail page
 * Small refactor to controller loading
+* Handle invalid tokens on the account page
 * Pubspec.yaml dependency updates
 
 ## [5.7.0] - 2022-01-29